+++ sql-plugin-api/src/main/scala/com/nvidia/spark/rapids/RapidsShuffleHeartbeatHandler.scala
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2021-2023, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nvidia.spark.rapids
+
+import org.apache.spark.storage.BlockManagerId
+
+
+/**
+ * This is the first message sent from the executor to the driver.
+ * @param id `BlockManagerId` for the executor
+ */
+case class RapidsExecutorStartupMsg(id: BlockManagerId)
+
+/**
+ * Executor heartbeat message.
+ * This gives the driver an opportunity to respond with `RapidsExecutorUpdateMsg`
+ */
+case class RapidsExecutorHeartbeatMsg(id: BlockManagerId)
+
+/**
+ * Driver response to an startup or heartbeat message, with new (to the peer) executors
+ * from the last heartbeat.
+ */
+case class RapidsExecutorUpdateMsg(ids: Array[BlockManagerId])
+
+trait RapidsShuffleHeartbeatHandler {
+  /** Called when a new peer is seen via heartbeats */
+  def addPeer(peer: BlockManagerId): Unit
+}
